refactor: formatdate 함수 utils폴더로 이동
