+++ src/app/(userpage)/_components/Transaction/TransactionItem.tsx
@@ -2,21 +2,12 @@ import { PropsWithChildren } from "react";
 
 import setMoneyUnitString from "@/utils/function/setMoneyUnitString";
 
-interface HeaderProps {
+import { formatDate } from "../../_utils/formatDate";
+
+export interface HeaderProps {
   firstLabel: "구매 희망가" | "최고 입찰가" | "낙찰가";
   secondLabel: "날짜" | "남은 시간" | "-";
 }
-
-const formatDate = (date?: Date | string) => {
-  if (date) {
-    const tempDate = new Date(date);
-    const year = tempDate.getFullYear().toString().substring(2);
-    const month = (tempDate.getMonth() + 1).toString().padStart(2, "0");
-    const day = tempDate.getDate().toString().padStart(2, "0");
-
-    return `${year}/${month}/${day}`;
-  }
-};
 
 export function TransactionHeader({ firstLabel, secondLabel }: HeaderProps) {
   return (
